requirements.txt
fastapi
uvicorn[standard]
mysql-connector-python
pymongo
redis
strawberry-graphql
python-dotenv
